requirements.txt
# --- Agent 核心框架 ---
# LangGraph 用于构建带循环（Cycle）和状态管理（State）的多智能体协作流
langgraph>=0.2.0
# LangChain 核心协议，支持 0.3.x 最新生态
langchain-core>=0.3.0
# 用于通过 OpenAI 兼容接口调用 DeepSeek-V3 或 Qwen2.5
langchain-openai>=0.2.0

# --- 环境管理与配置 ---
# 解析 .env 文件中的 API Key 和环境变量
python-dotenv>=1.0.0

# --- 前端展示层 ---
# 快速构建 Agent 决策链路的可视化看板
streamlit>=1.30.0

# --- 网络与辅助工具 ---
# 处理异步网络请求（若后续接入真实搜索 API 如 Tavily 时需要）
httpx>=0.27.0
tavily-python
langchain-community

# --- 后端接口层 ---
fastapi
uvicorn[standard]